+++ app/app.go
@@ -553,7 +553,7 @@ func New(
 		tracer,
 		evmS,
 		[]vm.PrecompiledContract{
-			cronosprecompiles.NewRelayerContract(app.IBCKeeper, appCodec, gasConfig),
+			cronosprecompiles.NewRelayerContract(app.IBCKeeper, appCodec, gasConfig, app.Logger()),
 			cronosprecompiles.NewIcaContract(&app.ICAAuthKeeper, &app.CronosKeeper, appCodec, gasConfig),
 		},
 		allKeys,